+++ src/providers/code-lens-actions.provider.ts
@@ -3,6 +3,7 @@ import * as vscode from "vscode";
 import { EXTENSION_NAME } from "../helpers/consts";
 import { MatchTypeChangeAnnotations } from "../helpers/types";
 import MyExtensionContext from "../helpers/my-extension.context";
+import { SettingsKeys } from "../scripts/settings";
 
 // const isTest = /^\s*(it|test)(?:\.(only|skip|fixme))?\s*\(\s*[\r\n]*\s*['"]/m;
 // const isSuite = /^\s*(describe|test\.describe)(?:\.(only|skip|fixme))?\s*\(\s*[\s\S]*?['"]/m;
@@ -19,7 +20,9 @@ function regexpIsSuite(annotations: string[]): RegExp {
 }
 
 export function matchTestAnnotations(document: vscode.TextDocument): vscode.CodeLens[] {
-  const testAnnotationsCodeLens = MyExtensionContext.instance.getWorkspaceValue("provideTestAnnotationsCodeLens");
+  const testAnnotationsCodeLens = MyExtensionContext.instance.getWorkspaceValue(
+    SettingsKeys.provideTestAnnotationsCodeLens
+  );
   if (!testAnnotationsCodeLens) {
     return [];
   }