+++ src/scripts/settings.ts
@@ -1,31 +1,39 @@
 import * as vscode from "vscode";
 import { TabViewCategory, ControlType, PwSettings } from "../helpers/types";
 
+export enum SettingsKeys {
+  reuseTerminal = "reuseTerminal",
+  verboseApiLogs = "verboseApiLogs",
+  neverOpenHtmlReport = "neverOpenHtmlReport",
+  provideTestAnnotationsCodeLens = "provideTestAnnotationsCodeLens",
+  instantExecute = "instantExecute",
+}
+
 export function getSettingsList(): PwSettings[] {
   const commandsList: PwSettings[] = [
     {
-      key: "reuseTerminal",
+      key: SettingsKeys.reuseTerminal,
       func: reuseTerminal,
       prettyName: vscode.l10n.t("Reuse Existing Terminal"),
       category: TabViewCategory.general,
       type: ControlType.checkbox,
     },
     {
-      key: "verboseApiLogs",
+      key: SettingsKeys.verboseApiLogs,
       func: verboseApiLogs,
       prettyName: vscode.l10n.t("Playwright Verbose API logs"),
       category: TabViewCategory.general,
       type: ControlType.checkbox,
     },
     {
-      key: "neverOpenHtmlReport",
+      key: SettingsKeys.neverOpenHtmlReport,
       func: neverOpenHtmlReport,
       prettyName: vscode.l10n.t("Never Open HTML Report"),
       category: TabViewCategory.general,
       type: ControlType.checkbox,
     },
     {
-      key: "provideTestAnnotationsCodeLens",
+      key: SettingsKeys.provideTestAnnotationsCodeLens,
       func: provideTestAnnotationsCodeLens,
       prettyName: vscode.l10n.t("Test Annotations CodeLens"),
       category: TabViewCategory.general,