+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "playwright-helpers" extension will be documented in 
 - feat: run tests from context menu
 - feat: possible versions to install
 - feat: add test annotations (skip, only, fixme) via CodeLens
+- feat: settings for test annotations
 
 ## [1.1.0] - 2024-09-14